BREAKING CHANGE: Change BatteryReadout::health return value to u8
There's no particular reason why one should allocate 64 bits for a
value that can only be <= 100.

As a bonus, ceil() the return value before finally casting to u8.

+++ src/traits.rs
@@ -74,7 +74,7 @@ impl BatteryReadout for MacOSBatteryReadout {
         Ok(BatteryState::Charging) //always charging.
     }
 
-    fn health(&self) -> Result<u64, ReadoutError>{
+    fn health(&self) -> Result<u8, ReadoutError>{
         //check the battery health...
         Ok(100) //totally healtyh
     }
@@ -95,7 +95,7 @@ pub trait BatteryReadout {
     fn status(&self) -> Result<BatteryState, ReadoutError>;
 
     /// This function is used for querying the current battery's health in percentage.
-    fn health(&self) -> Result<u64, ReadoutError>;
+    fn health(&self) -> Result<u8, ReadoutError>;
 }
 
 /**
